Adapted to code for compatibility with Webpack and Babel
Note:
Because JavaScript ES5 (transpiled by Babel from my Code in JavaScript ES6) doesn't support code that has circular dependencies, I had to refactor the _doAccordingToElementType method in MathElement.js, since it originally depended on the presence of Matrix.js and Vector.js, which again depend on MathElement.js. This forms a circular dependence.

To get rid of this problem, I had to introduce an attribute _type in MathElement.js, which stores a number, which indicates, if the current MathElement object is a number, a matrix or a vector.
This allows me to check the type of a MathElement object without knowing about the Matrix or Vector class. This is a horribly bad style, but was likely the easiest workaround.