Fixed a critical bug in Interpreter.js.
Interpreter.js used to interpret a math expression string from left to right, meaning that the left node of the root node always directly contains the first expression part and the right node always contains the entire rest of the tree. This causes miscalculations for several types of expressions:

E.g. the expression "1-1-1" used to be interpreted as "1-(1-1)" by the interpreter. However, this is a wrong interpretation, since 1-1-1=-1 but 1-(1-1)=0.

The issue was, that the interpreter interpreted the string from left to right. However, for a correct result the string must be interpreted from RIGHT TO LEFT!
Considering this change, "1-1-1" would be interpreted as "(1-1)-1" which is the correct interpretation, since (1-1)-1 = -1 = 1-1-1.

This commit fixes this issue.