Fixed critical bug in algorithm for solving homogeneous equation systems.
The bug caused that this._nonStepColumns in Matrix.js contained not all non-step-columns for some matrices.
This lead to the error that for some matrices the result was that there is no solution for the homogeneous equation system, even though there actually exists a solution.
E.g solvehom({1,2,3;2,2,4}) should have the solution [-1,-1,1] in R. However, due to the last column not being counted as a non-step-column, it was shown that the system does not have a solution.

This bug is now fixed by counting all columns, that were not touched during row-reducing to be counted as non-step-columns automatically.