Fixed bug that caused omitted multiplication operators before function operators not to be processed
E.g. 5rowreduce({1,2;3,4}) lead to a wrong result.

Additionally fixed a bug that caused the latex representation of the user output not to be displayed correctly in cases where a letter of an extended field is directly followed by a function operator.

E.g. arowreduce({0,1;a+1,a}) was not correctly rendered in Latex.

This has been fixed by just adding a space after each alpha, beta and iota character in Latex.