Use an epsilon when comparing real numbers
Rounding errors could lead to wrong results when row-reducing matrices over real numbers. For example the input
rowreduce({10,7,15;11,9,23;14,10,22})

leads to the WRONG result
1  0  0
0  1  0
0  0  1

whereas the correct result would be:
1  0 -2
0  1  5
0  0  0

The reason turned out the line 485 in Matrix.js:
if(this.value[row][column].value !== 0){ ...}

When this.value[row][column].value is not exactly 0 but a bit off due to rounding errors, the next column will be reduced which leads to a totally wrong result.

The solution is to introduce an Epsilon for comparing real numbers. This is what I implemented by introducing a new method equalsValue().

This bug only affected calculations with real numbers. Calculation with finite fields were not affected since they do not use floating point numbers.