Adapt style of matrix invertion to rest of code
I decided to make the following (stilistic) changes to the matrix inversion feature:
- Check if matrix has fitting dimensions in the inversion method itself, not in the query parser.
Mathematical checks should not be mixed with text parsing.

- Rename methods from invert to getMultiplicativeInverse
This convention is also used when calculating the multiplicative inverse of numbers.

- Don't introduce a new invert() operator but use multinverse() instead.
Reason: what we call 'inverting a matrix' actually means to calculate the *multiplicative* inverse of this matrix (one can also define an additive inverse of a matrix).
It is not desired to have two different operator keywords for the same concept. Therefore use multinverse operator for matrices as well.